change ssgからssrに変更&pathの変更に伴うファイルアーキテクト変更
